Feat: 카카오 로그인 POST 요청 완료 (#73)
2024-04-09 16:55
- KakaoAuthService.java: 카카오 로그인 서비스 로직을 분리
- MemberController.java: POST 요청 이후 발급된 JWT 토큰을 출력하도록 수정
- TODO: POST 요청 응답 JSON에서 JWT Payload decode 이후 email을 추출하여 DB에 삽입
